Trakt job: move call handling to base class.
